Extract some classes from Resolver.cs
