+++ Source/DafnyCore/Resolver/ExtremePredicateChecks_Visitor.cs
@@ -1,0 +1,56 @@
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Dafny;
+
+class ExtremePredicateChecks_Visitor : FindFriendlyCalls_Visitor {
+  readonly ExtremePredicate context;
+  public ExtremePredicateChecks_Visitor(ErrorReporter reporter, ExtremePredicate context)
+    : base(reporter, context is GreatestPredicate, context.KNat) {
+    Contract.Requires(reporter != null);
+    Contract.Requires(context != null);
+    this.context = context;
+  }
+  protected override bool VisitOneExpr(Expression expr, ref Resolver.CallingPosition cp) {
+    if (expr is FunctionCallExpr) {
+      var e = (FunctionCallExpr)expr;
+      if (ModuleDefinition.InSameSCC(context, e.Function)) {
+        // we're looking at a recursive call
+        if (!(context is LeastPredicate ? e.Function is LeastPredicate : e.Function is GreatestPredicate)) {
+          reporter.Error(MessageSource.Resolver, e, "a recursive call from a {0} can go only to other {0}s", context.WhatKind);
+        } else if (context.KNat != ((ExtremePredicate)e.Function).KNat) {
+          Resolver.KNatMismatchError(reporter, e.tok, context.Name, context.TypeOfK, ((ExtremePredicate)e.Function).TypeOfK);
+        } else if (cp != Resolver.CallingPosition.Positive) {
+          var msg = string.Format("a {0} can be called recursively only in positive positions", context.WhatKind);
+          if (ContinuityIsImportant && cp == Resolver.CallingPosition.Neither) {
+            // this may be inside an non-friendly quantifier
+            msg += string.Format(" and cannot sit inside an unbounded {0} quantifier", context is LeastPredicate ? "universal" : "existential");
+          } else {
+            // we don't care about the continuity restriction or
+            // the extreme-call is not inside an quantifier, so don't bother mentioning the part of existentials/universals in the error message
+          }
+          reporter.Error(MessageSource.Resolver, e, msg);
+        } else {
+          e.CoCall = FunctionCallExpr.CoCallResolution.Yes;
+          reporter.Info(MessageSource.Resolver, e.tok, e.Function.Name + "#[_k - 1]");
+        }
+      }
+      // do the sub-parts with cp := Neither
+      cp = Resolver.CallingPosition.Neither;
+      return true;
+    }
+    return base.VisitOneExpr(expr, ref cp);
+  }
+  protected override bool VisitOneStmt(Statement stmt, ref Resolver.CallingPosition st) {
+    if (stmt is CallStmt) {
+      var s = (CallStmt)stmt;
+      if (ModuleDefinition.InSameSCC(context, s.Method)) {
+        // we're looking at a recursive call
+        reporter.Error(MessageSource.Resolver, stmt.Tok, "a recursive call from a {0} can go only to other {0}s", context.WhatKind);
+      }
+      // do the sub-parts with the same "cp"
+      return true;
+    } else {
+      return base.VisitOneStmt(stmt, ref st);
+    }
+  }
+}
