+++ Source/DafnyCore/Resolver/Resolver.cs
@@ -61,120 +61,6 @@ namespace Microsoft.Dafny {
       return freshTempVarName;
     }
 
-    interface IAmbiguousThing<Thing> {
-      /// <summary>
-      /// Returns a plural number of non-null Thing's
-      /// </summary>
-      ISet<Thing> Pool { get; }
-    }
-
-    class AmbiguousThingHelper<Thing> where Thing : class {
-      public static Thing Create(ModuleDefinition m, Thing a, Thing b, IEqualityComparer<Thing> eq, out ISet<Thing> s) {
-        Contract.Requires(a != null);
-        Contract.Requires(b != null);
-        Contract.Requires(eq != null);
-        Contract.Ensures(Contract.Result<Thing>() != null ||
-                         Contract.ValueAtReturn(out s) != null || 2 <= Contract.ValueAtReturn(out s).Count);
-        s = null;
-        if (eq.Equals(a, b)) {
-          return a;
-        }
-
-        ISet<Thing> sa = a is IAmbiguousThing<Thing> ? ((IAmbiguousThing<Thing>)a).Pool : new HashSet<Thing>() { a };
-        ISet<Thing> sb = b is IAmbiguousThing<Thing> ? ((IAmbiguousThing<Thing>)b).Pool : new HashSet<Thing>() { b };
-        var union = new HashSet<Thing>(sa.Union(sb, eq));
-        if (sa.Count == union.Count) {
-          // sb is a subset of sa
-          return a;
-        } else if (sb.Count == union.Count) {
-          // sa is a subset of sb
-          return b;
-        } else {
-          s = union;
-          Contract.Assert(2 <= s.Count);
-          return null;
-        }
-      }
-
-      public static string Name(ISet<Thing> s, Func<Thing, string> name) {
-        Contract.Requires(s != null);
-        Contract.Requires(s.Count != 0);
-        string nm = null;
-        foreach (var thing in s) {
-          string n = name(thing);
-          if (nm == null) {
-            nm = n;
-          } else {
-            nm += "/" + n;
-          }
-        }
-
-        return nm;
-      }
-
-      public static string ModuleNames(IAmbiguousThing<Thing> amb, Func<Thing, string> moduleName) {
-        Contract.Requires(amb != null);
-        Contract.Ensures(Contract.Result<string>() != null);
-        string nm = null;
-        foreach (var d in amb.Pool) {
-          if (nm == null) {
-            nm = moduleName(d);
-          } else {
-            nm += ", " + moduleName(d);
-          }
-        }
-
-        return nm;
-      }
-    }
-
-    public class AmbiguousTopLevelDecl : TopLevelDecl, IAmbiguousThing<TopLevelDecl> // only used with "classes"
-    {
-      public static TopLevelDecl Create(ModuleDefinition m, TopLevelDecl a, TopLevelDecl b) {
-        var t = AmbiguousThingHelper<TopLevelDecl>.Create(m, a, b, new Eq(), out var s);
-        return t ?? new AmbiguousTopLevelDecl(m, AmbiguousThingHelper<TopLevelDecl>.Name(s, tld => tld.Name), s);
-      }
-
-      class Eq : IEqualityComparer<TopLevelDecl> {
-        public bool Equals(TopLevelDecl d0, TopLevelDecl d1) {
-          // We'd like to resolve any AliasModuleDecl to whatever module they refer to.
-          // It seems that the only way to do that is to look at alias.Signature.ModuleDef,
-          // but that is a ModuleDefinition, which is not a TopLevelDecl.  Therefore, we
-          // convert to a ModuleDefinition anything that might refer to something that an
-          // AliasModuleDecl can refer to; this is AliasModuleDecl and LiteralModuleDecl.
-          object a = d0 is ModuleDecl ? ((ModuleDecl)d0).Dereference() : d0;
-          object b = d1 is ModuleDecl ? ((ModuleDecl)d1).Dereference() : d1;
-          return a == b;
-        }
-
-        public int GetHashCode(TopLevelDecl d) {
-          object a = d is ModuleDecl ? ((ModuleDecl)d).Dereference() : d;
-          return a.GetHashCode();
-        }
-      }
-
-      public override string WhatKind {
-        get { return Pool.First().WhatKind; }
-      }
-
-      readonly ISet<TopLevelDecl> Pool = new HashSet<TopLevelDecl>();
-
-      ISet<TopLevelDecl> IAmbiguousThing<TopLevelDecl>.Pool {
-        get { return Pool; }
-      }
-
-      private AmbiguousTopLevelDecl(ModuleDefinition m, string name, ISet<TopLevelDecl> pool)
-        : base(pool.First().RangeToken, new Name(pool.First().RangeToken, name), m, new List<TypeParameter>(), null, false) {
-        Contract.Requires(name != null);
-        Contract.Requires(pool != null && 2 <= pool.Count);
-        Pool = pool;
-      }
-
-      public string ModuleNames() {
-        return AmbiguousThingHelper<TopLevelDecl>.ModuleNames(this, d => d.EnclosingModuleDefinition.Name);
-      }
-    }
-
     class AmbiguousMemberDecl : MemberDecl, IAmbiguousThing<MemberDecl> // only used with "classes"
     {
       public static MemberDecl Create(ModuleDefinition m, MemberDecl a, MemberDecl b) {
@@ -3775,19 +3661,11 @@ namespace Microsoft.Dafny {
     #endregion
 
     // ------------------------------------------------------------------------------------------------------
-    // ----- FuelAdjustmentChecks ---------------------------------------------------------------------------
-    // ------------------------------------------------------------------------------------------------------
-    #region FuelAdjustmentChecks
-
-
-    #endregion FuelAdjustmentChecks
-
-    // ------------------------------------------------------------------------------------------------------
     // ----- ExtremePredicateChecks -------------------------------------------------------------------------
     // ------------------------------------------------------------------------------------------------------
     #region ExtremePredicateChecks
-    enum CallingPosition { Positive, Negative, Neither }
-    static CallingPosition Invert(CallingPosition cp) {
+    public enum CallingPosition { Positive, Negative, Neither }
+    public static CallingPosition Invert(CallingPosition cp) {
       switch (cp) {
         case CallingPosition.Positive: return CallingPosition.Negative;
         case CallingPosition.Negative: return CallingPosition.Positive;
@@ -3795,161 +3673,13 @@ namespace Microsoft.Dafny {
       }
     }
 
-    class FindFriendlyCalls_Visitor : ResolverTopDownVisitor<CallingPosition> {
-      public readonly bool IsCoContext;
-      public readonly bool ContinuityIsImportant;
-      public FindFriendlyCalls_Visitor(ErrorReporter reporter, bool co, bool continuityIsImportant)
-        : base(reporter) {
-        Contract.Requires(reporter != null);
-        this.IsCoContext = co;
-        this.ContinuityIsImportant = continuityIsImportant;
-      }
-
-      protected override bool VisitOneExpr(Expression expr, ref CallingPosition cp) {
-        if (expr is UnaryOpExpr) {
-          var e = (UnaryOpExpr)expr;
-          if (e.Op == UnaryOpExpr.Opcode.Not) {
-            // for the sub-parts, use Invert(cp)
-            cp = Invert(cp);
-            return true;
-          }
-        } else if (expr is BinaryExpr) {
-          var e = (BinaryExpr)expr;
-          switch (e.ResolvedOp) {
-            case BinaryExpr.ResolvedOpcode.And:
-            case BinaryExpr.ResolvedOpcode.Or:
-              return true;  // do the sub-parts with the same "cp"
-            case BinaryExpr.ResolvedOpcode.Imp:
-              Visit(e.E0, Invert(cp));
-              Visit(e.E1, cp);
-              return false;  // don't recurse (again) on the sub-parts
-            default:
-              break;
-          }
-        } else if (expr is NestedMatchExpr) {
-          var e = (NestedMatchExpr)expr;
-          Visit(e.Source, CallingPosition.Neither);
-          var theCp = cp;
-          e.Cases.Iter(kase => Visit(kase.Body, theCp));
-          return false;
-        } else if (expr is MatchExpr) {
-          var e = (MatchExpr)expr;
-          Visit(e.Source, CallingPosition.Neither);
-          var theCp = cp;
-          e.Cases.Iter(kase => Visit(kase.Body, theCp));
-          return false;
-        } else if (expr is ITEExpr) {
-          var e = (ITEExpr)expr;
-          Visit(e.Test, CallingPosition.Neither);
-          Visit(e.Thn, cp);
-          Visit(e.Els, cp);
-          return false;
-        } else if (expr is LetExpr) {
-          var e = (LetExpr)expr;
-          foreach (var rhs in e.RHSs) {
-            Visit(rhs, CallingPosition.Neither);
-          }
-          var cpBody = cp;
-          if (!e.Exact) {
-            // a let-such-that expression introduces an existential that may depend on the _k in a least/greatest predicate, so we disallow recursive calls in the body of the let-such-that
-            if (IsCoContext && cp == CallingPosition.Positive) {
-              cpBody = CallingPosition.Neither;
-            } else if (!IsCoContext && cp == CallingPosition.Negative) {
-              cpBody = CallingPosition.Neither;
-            }
-          }
-          Visit(e.Body, cpBody);
-          return false;
-        } else if (expr is QuantifierExpr) {
-          var e = (QuantifierExpr)expr;
-          Contract.Assert(e.SplitQuantifier == null); // No split quantifiers during resolution
-          var cpos = IsCoContext ? cp : Invert(cp);
-          if (ContinuityIsImportant) {
-            if ((cpos == CallingPosition.Positive && e is ExistsExpr) || (cpos == CallingPosition.Negative && e is ForallExpr)) {
-              if (e.Bounds.Exists(bnd => bnd == null || (bnd.Virtues & ComprehensionExpr.BoundedPool.PoolVirtues.Finite) == 0)) {
-                // To ensure continuity of extreme predicates, don't allow calls under an existential (resp. universal) quantifier
-                // for greatest (resp. least) predicates).
-                cp = CallingPosition.Neither;
-              }
-            }
-          }
-          Visit(e.LogicalBody(), cp);
-          return false;
-        } else if (expr is StmtExpr) {
-          var e = (StmtExpr)expr;
-          Visit(e.E, cp);
-          Visit(e.S, CallingPosition.Neither);
-          return false;
-        } else if (expr is ConcreteSyntaxExpression) {
-          // do the sub-parts with the same "cp"
-          return true;
-        }
-        // do the sub-parts with cp := Neither
-        cp = CallingPosition.Neither;
-        return true;
-      }
-    }
-
-    static void KNatMismatchError(ErrorReporter reporter, IToken tok, string contextName, ExtremePredicate.KType contextK, ExtremePredicate.KType calleeK) {
+    public static void KNatMismatchError(ErrorReporter reporter, IToken tok, string contextName, ExtremePredicate.KType contextK, ExtremePredicate.KType calleeK) {
       var hint = contextK == ExtremePredicate.KType.Unspecified ? string.Format(" (perhaps try declaring '{0}' as '{0}[nat]')", contextName) : "";
       reporter.Error(MessageSource.Resolver, tok,
         "this call does not type check, because the context uses a _k parameter of type {0} whereas the callee uses a _k parameter of type {1}{2}",
         contextK == ExtremePredicate.KType.Nat ? "nat" : "ORDINAL",
         calleeK == ExtremePredicate.KType.Nat ? "nat" : "ORDINAL",
         hint);
-    }
-
-    class ExtremePredicateChecks_Visitor : FindFriendlyCalls_Visitor {
-      readonly ExtremePredicate context;
-      public ExtremePredicateChecks_Visitor(ErrorReporter reporter, ExtremePredicate context)
-        : base(reporter, context is GreatestPredicate, context.KNat) {
-        Contract.Requires(reporter != null);
-        Contract.Requires(context != null);
-        this.context = context;
-      }
-      protected override bool VisitOneExpr(Expression expr, ref CallingPosition cp) {
-        if (expr is FunctionCallExpr) {
-          var e = (FunctionCallExpr)expr;
-          if (ModuleDefinition.InSameSCC(context, e.Function)) {
-            // we're looking at a recursive call
-            if (!(context is LeastPredicate ? e.Function is LeastPredicate : e.Function is GreatestPredicate)) {
-              reporter.Error(MessageSource.Resolver, e, "a recursive call from a {0} can go only to other {0}s", context.WhatKind);
-            } else if (context.KNat != ((ExtremePredicate)e.Function).KNat) {
-              Resolver.KNatMismatchError(reporter, e.tok, context.Name, context.TypeOfK, ((ExtremePredicate)e.Function).TypeOfK);
-            } else if (cp != CallingPosition.Positive) {
-              var msg = string.Format("a {0} can be called recursively only in positive positions", context.WhatKind);
-              if (ContinuityIsImportant && cp == CallingPosition.Neither) {
-                // this may be inside an non-friendly quantifier
-                msg += string.Format(" and cannot sit inside an unbounded {0} quantifier", context is LeastPredicate ? "universal" : "existential");
-              } else {
-                // we don't care about the continuity restriction or
-                // the extreme-call is not inside an quantifier, so don't bother mentioning the part of existentials/universals in the error message
-              }
-              reporter.Error(MessageSource.Resolver, e, msg);
-            } else {
-              e.CoCall = FunctionCallExpr.CoCallResolution.Yes;
-              reporter.Info(MessageSource.Resolver, e.tok, e.Function.Name + "#[_k - 1]");
-            }
-          }
-          // do the sub-parts with cp := Neither
-          cp = CallingPosition.Neither;
-          return true;
-        }
-        return base.VisitOneExpr(expr, ref cp);
-      }
-      protected override bool VisitOneStmt(Statement stmt, ref CallingPosition st) {
-        if (stmt is CallStmt) {
-          var s = (CallStmt)stmt;
-          if (ModuleDefinition.InSameSCC(context, s.Method)) {
-            // we're looking at a recursive call
-            reporter.Error(MessageSource.Resolver, stmt.Tok, "a recursive call from a {0} can go only to other {0}s", context.WhatKind);
-          }
-          // do the sub-parts with the same "cp"
-          return true;
-        } else {
-          return base.VisitOneStmt(stmt, ref st);
-        }
-      }
     }
 
     void ExtremePredicateChecks(Expression expr, ExtremePredicate context, CallingPosition cp) {
@@ -6554,48 +6284,6 @@ namespace Microsoft.Dafny {
       var visitor = new CollectFriendlyCallsInSpec_Visitor(reporter, friendlyCalls, co, context);
       visitor.Visit(expr, position ? CallingPosition.Positive : CallingPosition.Negative);
     }
-
-    class CollectFriendlyCallsInSpec_Visitor : FindFriendlyCalls_Visitor {
-      readonly ISet<Expression> friendlyCalls;
-      readonly ExtremeLemma Context;
-      public CollectFriendlyCallsInSpec_Visitor(ErrorReporter reporter, ISet<Expression> friendlyCalls, bool co, ExtremeLemma context)
-        : base(reporter, co, context.KNat) {
-        Contract.Requires(reporter != null);
-        Contract.Requires(friendlyCalls != null);
-        Contract.Requires(context != null);
-        this.friendlyCalls = friendlyCalls;
-        this.Context = context;
-      }
-      protected override bool VisitOneExpr(Expression expr, ref CallingPosition cp) {
-        if (cp == CallingPosition.Neither) {
-          // no friendly calls in "expr"
-          return false;  // don't recurse into subexpressions
-        }
-        if (expr is FunctionCallExpr) {
-          if (cp == CallingPosition.Positive) {
-            var fexp = (FunctionCallExpr)expr;
-            if (IsCoContext ? fexp.Function is GreatestPredicate : fexp.Function is LeastPredicate) {
-              if (Context.KNat != ((ExtremePredicate)fexp.Function).KNat) {
-                Resolver.KNatMismatchError(reporter, expr.tok, Context.Name, Context.TypeOfK, ((ExtremePredicate)fexp.Function).TypeOfK);
-              } else {
-                friendlyCalls.Add(fexp);
-              }
-            }
-          }
-          return false;  // don't explore subexpressions any further
-        } else if (expr is BinaryExpr && IsCoContext) {
-          var bin = (BinaryExpr)expr;
-          if (cp == CallingPosition.Positive && bin.ResolvedOp == BinaryExpr.ResolvedOpcode.EqCommon && bin.E0.Type.IsCoDatatype) {
-            friendlyCalls.Add(bin);
-            return false;  // don't explore subexpressions any further
-          } else if (cp == CallingPosition.Negative && bin.ResolvedOp == BinaryExpr.ResolvedOpcode.NeqCommon && bin.E0.Type.IsCoDatatype) {
-            friendlyCalls.Add(bin);
-            return false;  // don't explore subexpressions any further
-          }
-        }
-        return base.VisitOneExpr(expr, ref cp);
-      }
-    }
   }
 
   abstract class ResolverTopDownVisitor<T> : TopDownVisitor<T> {
@@ -6603,279 +6291,6 @@ namespace Microsoft.Dafny {
     public ResolverTopDownVisitor(ErrorReporter reporter) {
       Contract.Requires(reporter != null);
       this.reporter = reporter;
-    }
-  }
-
-  class CoCallResolution {
-    readonly Function currentFunction;
-    readonly bool dealsWithCodatatypes;
-    public bool HasIntraClusterCallsInDestructiveContexts = false;
-    public readonly List<CoCallInfo> FinalCandidates = new List<CoCallInfo>();
-
-    public CoCallResolution(Function currentFunction, bool dealsWithCodatatypes) {
-      Contract.Requires(currentFunction != null);
-      this.currentFunction = currentFunction;
-      this.dealsWithCodatatypes = dealsWithCodatatypes;
-    }
-
-    /// <summary>
-    /// Determines which calls in "expr" can be considered to be co-calls, which co-constructor
-    /// invocations host such co-calls, and which destructor operations are not allowed.
-    /// Also records whether or not there are any intra-cluster calls in a destructive context.
-    /// Assumes "expr" to have been successfully resolved.
-    /// </summary>
-    public void CheckCoCalls(Expression expr) {
-      Contract.Requires(expr != null);
-      CheckCoCalls(expr, 0, null, FinalCandidates);
-    }
-
-    public struct CoCallInfo {
-      public readonly FunctionCallExpr CandidateCall;
-      public readonly DatatypeValue EnclosingCoConstructor;
-      public CoCallInfo(FunctionCallExpr candidateCall, DatatypeValue enclosingCoConstructor) {
-        Contract.Requires(candidateCall != null);
-        Contract.Requires(enclosingCoConstructor != null);
-        CandidateCall = candidateCall;
-        EnclosingCoConstructor = enclosingCoConstructor;
-      }
-    }
-
-    /// <summary>
-    /// Recursively goes through the entire "expr".  Every call within the same recursive cluster is a potential
-    /// co-call.  If the call is determined not to be a co-recursive call, then its .CoCall field is filled in;
-    /// if the situation deals with co-datatypes, then one of the NoBecause... values is chosen (rather
-    /// than just No), so that any error message that may later be produced when trying to prove termination of the
-    /// recursive call can include a note pointing out that the call was not selected to be a co-call.
-    /// If the call looks like it is guarded, then it is added to the list "coCandicates", so that a later analysis
-    /// can either set all of those .CoCall fields to Yes or to NoBecauseRecursiveCallsInDestructiveContext, depending
-    /// on other intra-cluster calls.
-    /// The "destructionLevel" indicates how many pending co-destructors the context has.  It may be infinity (int.MaxValue)
-    /// if the enclosing context has no easy way of controlling the uses of "expr" (for example, if the enclosing context
-    /// passes "expr" to a function or binds "expr" to a variable).  It is never negative -- excess co-constructors are
-    /// not considered an asset, and any immediately enclosing co-constructor is passed in as a non-null "coContext" anyway.
-    /// "coContext" is non-null if the immediate context is a co-constructor.
-    /// </summary>
-    void CheckCoCalls(Expression expr, int destructionLevel, DatatypeValue coContext, List<CoCallInfo> coCandidates, Function functionYouMayWishWereAbstemious = null) {
-      Contract.Requires(expr != null);
-
-      Contract.Requires(0 <= destructionLevel);
-      Contract.Requires(coCandidates != null);
-      expr = expr.Resolved;
-      if (expr is DatatypeValue) {
-        var e = (DatatypeValue)expr;
-        if (e.Ctor.EnclosingDatatype is CoDatatypeDecl) {
-          int dl = destructionLevel == int.MaxValue ? int.MaxValue : destructionLevel == 0 ? 0 : destructionLevel - 1;
-          foreach (var arg in e.Arguments) {
-            CheckCoCalls(arg, dl, e, coCandidates);
-          }
-          return;
-        }
-      } else if (expr is MemberSelectExpr) {
-        var e = (MemberSelectExpr)expr;
-        if (e.Member.EnclosingClass is CoDatatypeDecl) {
-          int dl = destructionLevel == int.MaxValue ? int.MaxValue : destructionLevel + 1;
-          CheckCoCalls(e.Obj, dl, coContext, coCandidates);
-          return;
-        }
-      } else if (expr is BinaryExpr) {
-        var e = (BinaryExpr)expr;
-        if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.EqCommon || e.ResolvedOp == BinaryExpr.ResolvedOpcode.NeqCommon) {
-          // Equality and disequality (for any type that may contain a co-datatype) are as destructive as can be--in essence,
-          // they destruct the values indefinitely--so don't allow any co-recursive calls in the operands.
-          CheckCoCalls(e.E0, int.MaxValue, null, coCandidates);
-          CheckCoCalls(e.E1, int.MaxValue, null, coCandidates);
-          return;
-        }
-      } else if (expr is TernaryExpr) {
-        var e = (TernaryExpr)expr;
-        if (e.Op == TernaryExpr.Opcode.PrefixEqOp || e.Op == TernaryExpr.Opcode.PrefixNeqOp) {
-          // Prefix equality and disequality (for any type that may contain a co-datatype) are destructive.
-          CheckCoCalls(e.E0, int.MaxValue, null, coCandidates);
-          CheckCoCalls(e.E1, int.MaxValue, null, coCandidates);
-          CheckCoCalls(e.E2, int.MaxValue, null, coCandidates);
-          return;
-        }
-      } else if (expr is NestedMatchExpr) {
-        var e = (NestedMatchExpr)expr;
-        foreach (var child in e.SubExpressions) {
-          CheckCoCalls(child, destructionLevel, coContext, coCandidates);
-        }
-      } else if (expr is MatchExpr) {
-        var e = (MatchExpr)expr;
-        CheckCoCalls(e.Source, int.MaxValue, null, coCandidates);
-        foreach (var kase in e.Cases) {
-          CheckCoCalls(kase.Body, destructionLevel, coContext, coCandidates);
-        }
-        return;
-      } else if (expr is ITEExpr) {
-        var e = (ITEExpr)expr;
-        CheckCoCalls(e.Test, int.MaxValue, null, coCandidates);
-        CheckCoCalls(e.Thn, destructionLevel, coContext, coCandidates);
-        CheckCoCalls(e.Els, destructionLevel, coContext, coCandidates);
-        return;
-      } else if (expr is FunctionCallExpr) {
-        var e = (FunctionCallExpr)expr;
-        // First, consider the arguments of the call, making sure that they do not include calls within the recursive cluster,
-        // unless the callee is abstemious.
-        var abstemious = true;
-        if (!Attributes.ContainsBool(e.Function.Attributes, "abstemious", ref abstemious)) {
-          abstemious = false;
-        }
-        Contract.Assert(e.Args.Count == e.Function.Formals.Count);
-        for (var i = 0; i < e.Args.Count; i++) {
-          var arg = e.Args[i];
-          if (!e.Function.Formals[i].Type.IsCoDatatype) {
-            CheckCoCalls(arg, int.MaxValue, null, coCandidates);
-          } else if (abstemious) {
-            CheckCoCalls(arg, 0, coContext, coCandidates);
-          } else {
-            // don't you wish the callee were abstemious
-            CheckCoCalls(arg, int.MaxValue, null, coCandidates, e.Function);
-          }
-        }
-        // Second, investigate the possibility that this call itself may be a candidate co-call
-        if (e.Name != "requires" && ModuleDefinition.InSameSCC(currentFunction, e.Function)) {
-          // This call goes to another function in the same recursive cluster
-          if (destructionLevel != 0 && GuaranteedCoCtors(e.Function) <= destructionLevel) {
-            // a potentially destructive context
-            HasIntraClusterCallsInDestructiveContexts = true;  // this says we found an intra-cluster call unsuitable for recursion, if there were any co-recursive calls
-            if (!dealsWithCodatatypes) {
-              e.CoCall = FunctionCallExpr.CoCallResolution.No;
-            } else {
-              e.CoCall = FunctionCallExpr.CoCallResolution.NoBecauseRecursiveCallsAreNotAllowedInThisContext;
-              if (functionYouMayWishWereAbstemious != null) {
-                e.CoCallHint = string.Format("perhaps try declaring function '{0}' with '{{:abstemious}}'", functionYouMayWishWereAbstemious.Name);
-              }
-            }
-          } else if (coContext == null) {
-            // no immediately enclosing co-constructor
-            if (!dealsWithCodatatypes) {
-              e.CoCall = FunctionCallExpr.CoCallResolution.No;
-            } else {
-              e.CoCall = FunctionCallExpr.CoCallResolution.NoBecauseIsNotGuarded;
-            }
-          } else if (e.Function.Reads.Count != 0) {
-            // this call is disqualified from being a co-call, because of side effects
-            if (!dealsWithCodatatypes) {
-              e.CoCall = FunctionCallExpr.CoCallResolution.No;
-            } else {
-              e.CoCall = FunctionCallExpr.CoCallResolution.NoBecauseFunctionHasSideEffects;
-            }
-          } else if (e.Function.Ens.Count != 0) {
-            // this call is disqualified from being a co-call, because it has a postcondition
-            // (a postcondition could be allowed, as long as it does not get to be used with
-            // co-recursive calls, because that could be unsound; for example, consider
-            // "ensures false")
-            if (!dealsWithCodatatypes) {
-              e.CoCall = FunctionCallExpr.CoCallResolution.No;
-            } else {
-              e.CoCall = FunctionCallExpr.CoCallResolution.NoBecauseFunctionHasPostcondition;
-            }
-          } else {
-            // e.CoCall is not filled in here, but will be filled in when the list of candidates are processed
-            coCandidates.Add(new CoCallInfo(e, coContext));
-          }
-        }
-        return;
-      } else if (expr is LambdaExpr) {
-        var e = (LambdaExpr)expr;
-        CheckCoCalls(e.Term, destructionLevel, coContext, coCandidates);
-        if (e.Range != null) {
-          CheckCoCalls(e.Range, int.MaxValue, null, coCandidates);
-        }
-        foreach (var read in e.Reads) {
-          CheckCoCalls(read.E, int.MaxValue, null, coCandidates);
-        }
-        return;
-      } else if (expr is MapComprehension) {
-        var e = (MapComprehension)expr;
-        foreach (var ee in Attributes.SubExpressions(e.Attributes)) {
-          CheckCoCalls(ee, int.MaxValue, null, coCandidates);
-        }
-        if (e.Range != null) {
-          CheckCoCalls(e.Range, int.MaxValue, null, coCandidates);
-        }
-        // allow co-calls in the term
-        if (e.TermLeft != null) {
-          CheckCoCalls(e.TermLeft, destructionLevel, coContext, coCandidates);
-        }
-        CheckCoCalls(e.Term, destructionLevel, coContext, coCandidates);
-        return;
-      } else if (expr is OldExpr) {
-        var e = (OldExpr)expr;
-        // here, "coContext" is passed along (the use of "old" says this must be ghost code, so the compiler does not need to handle this case)
-        CheckCoCalls(e.E, destructionLevel, coContext, coCandidates);
-        return;
-      } else if (expr is LetExpr) {
-        var e = (LetExpr)expr;
-        foreach (var rhs in e.RHSs) {
-          CheckCoCalls(rhs, int.MaxValue, null, coCandidates);
-        }
-        CheckCoCalls(e.Body, destructionLevel, coContext, coCandidates);
-        return;
-      } else if (expr is ApplyExpr) {
-        var e = (ApplyExpr)expr;
-        CheckCoCalls(e.Function, int.MaxValue, null, coCandidates);
-        foreach (var ee in e.Args) {
-          CheckCoCalls(ee, destructionLevel, null, coCandidates);
-        }
-        return;
-      }
-
-      // Default handling:
-      foreach (var ee in expr.SubExpressions) {
-        CheckCoCalls(ee, destructionLevel, null, coCandidates);
-      }
-    }
-
-    public static int GuaranteedCoCtors(Function function) {
-      Contract.Requires(function != null);
-      return function.Body != null ? GuaranteedCoCtorsAux(function.Body) : 0;
-    }
-
-    private static int GuaranteedCoCtorsAux(Expression expr) {
-      Contract.Requires(expr != null);
-      expr = expr.Resolved;
-      if (expr is DatatypeValue) {
-        var e = (DatatypeValue)expr;
-        if (e.Ctor.EnclosingDatatype is CoDatatypeDecl) {
-          var minOfArgs = int.MaxValue;  // int.MaxValue means: not yet encountered a formal whose type is a co-datatype
-          Contract.Assert(e.Arguments.Count == e.Ctor.Formals.Count);
-          for (var i = 0; i < e.Arguments.Count; i++) {
-            if (e.Ctor.Formals[i].Type.IsCoDatatype) {
-              var n = GuaranteedCoCtorsAux(e.Arguments[i]);
-              minOfArgs = Math.Min(minOfArgs, n);
-            }
-          }
-          return minOfArgs == int.MaxValue ? 1 : 1 + minOfArgs;
-        }
-      } else if (expr is ITEExpr) {
-        var e = (ITEExpr)expr;
-        var thn = GuaranteedCoCtorsAux(e.Thn);
-        var els = GuaranteedCoCtorsAux(e.Els);
-        return thn < els ? thn : els;
-      } else if (expr is NestedMatchExpr nestedMatchExpr) {
-        var childValues = nestedMatchExpr.Cases.Select(child => GuaranteedCoCtorsAux(child.Body)).ToList();
-        return childValues.Any() ? childValues.Min() : 0;
-      } else if (expr is MatchExpr) {
-        var e = (MatchExpr)expr;
-        var min = int.MaxValue;
-        foreach (var kase in e.Cases) {
-          var n = GuaranteedCoCtorsAux(kase.Body);
-          min = Math.Min(min, n);
-        }
-        return min == int.MaxValue ? 0 : min;
-      } else if (expr is LetExpr) {
-        var e = (LetExpr)expr;
-        return GuaranteedCoCtorsAux(e.Body);
-      } else if (expr is IdentifierExpr) {
-        var e = (IdentifierExpr)expr;
-        if (e.Type.IsCoDatatype && e.Var is Formal) {
-          // even though this is not a co-constructor, count this as 1, since that's what we would have done if it were, e.g., "Cons(s.head, s.tail)" instead of "s"
-          return 1;
-        }
-      }
-      return 0;
     }
   }
 }
